SD Refund 로직 수정 및 자료형 수정

+++ src/main/java/com/youthcase/orderflow/pr/controller/InventoryController.java
@@ -16,7 +16,7 @@ public class InventoryController {
     private final InventoryService inv;
 
     //조회
-    record AvailDto(String gtin, int available) {}
+    record AvailDto(String gtin, Long available) {}
     @GetMapping("/inventory")
     public AvailDto available(@RequestParam String gtin){
         return new AvailDto(gtin, inv.getAvailable(gtin));
