+++ src/main/java/com/youthcase/orderflow/pr/dto/SelectionItemDto.java
@@ -3,4 +3,4 @@ package com.youthcase.orderflow.pr.dto;
 import java.time.LocalDate;
 
 public record SelectionItemDto(
-        String gtin, Integer quantity, LocalDate orderDate) {}
+        String gtin, Long quantity, LocalDate orderDate) {}
