+++ src/main/java/com/youthcase/orderflow/sd/sdPayment/dto/VerifyResponse.java
@@ -2,6 +2,7 @@ package com.youthcase.orderflow.sd.sdPayment.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 @Builder
@@ -10,7 +11,7 @@ public record VerifyResponse(
         String merchantUid,
         String transactionNo,     // 내부 승인번호 (CARD, CASH)
         String status,          // "paid", "ready", "failed" 등
-        Integer amount,         // 결제 금액
+        BigDecimal amount,         // 결제 금액
         String currency,        // KRW
         String pgProvider,      // kakaopay / tosspayments ...
         String payMethod,       // card, vbank ...
